+++ apps/io-session-manager/src/services/redis-session-storage.ts
@@ -323,6 +323,34 @@ const readSessionInfoKeys =
       arrayStringReplyAsync,
     );
 
+/**
+ * Retrieves the session info keys related to a user fiscal code.
+ * If there are no session info keys, it returns an empty array,
+ * in order to trigger the normal flow for the case of no active session for the user.
+ * 
+ * Otherwise it returns the error.
+ * @param fiscalCode the fiscal code of the user
+ * @returns an array of session info keys, possibly empty, or an error
+ */
+export const retrieveSessionInfoKeys =
+  (redisClientSelector: RedisClientSelectorType) =>
+  (fiscalCode: FiscalCode): TE.TaskEither<Error, ReadonlyArray<string>> =>
+    pipe(
+      TE.fromTask(() => readSessionInfoKeys(redisClientSelector)(fiscalCode)),
+      TE.chainEitherKW(
+        E.orElseW((err) =>
+          err === RedisRepo.sessionNotFoundError
+            ? E.right(ROA.empty)
+            : E.left(err),
+        ),
+      ),
+    );
+
+const removePrefixFromSessionInfoKey = (key: string): string =>
+  key.replace(RedisRepo.sessionInfoKeyPrefix, "");
+
+export const removePrefixFromSessionInfoKeys = (keys: ReadonlyArray<string>): ReadonlyArray<string> =>
+  keys.map(removePrefixFromSessionInfoKey);
 // ---------------------------------
 // User tokens
 // ---------------------------------